改为多个bin文件入口
