+++ user/apps/user-manage/Cargo.toml
@@ -5,31 +5,31 @@ edition = "2021"
 
 [[bin]]
 name = "useradd"
-path = "src/main.rs"
+path = "src/cmd/useradd.rs"
 
 [[bin]]
 name = "userdel"
-path = "src/main.rs"
+path = "src/cmd/userdel.rs"
 
 [[bin]]
 name = "usermod"
-path = "src/main.rs"
+path = "src/cmd/usermod.rs"
 
 [[bin]]
 name = "passwd"
-path = "src/main.rs"
+path = "src/cmd/passwd.rs"
 
 [[bin]]
 name = "groupadd"
-path = "src/main.rs"
+path = "src/cmd/groupadd.rs"
 
 [[bin]]
 name = "groupdel"
-path = "src/main.rs"
+path = "src/cmd/groupdel.rs"
 
 [[bin]]
 name = "groupmod"
-path = "src/main.rs"
+path = "src/cmd/groupmod.rs"
 
 [dependencies]
 libc = "0.2.153"
