+++ user/apps/user-manage/src/cmd/mod.rs
@@ -1,0 +1,7 @@
+mod groupadd;
+mod groupdel;
+mod groupmod;
+mod passwd;
+mod useradd;
+mod userdel;
+mod usermod;
