+++ user/apps/user-manage/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod check;
+pub mod cmd;
 pub mod error;
 pub mod executor;
 pub mod parser;
